fix(generate): infer float8 for |/ and ||/ prefix operators

+++ packages/generate/src/generate/generate.operators.test.ts
@@ -338,14 +338,14 @@ test(`'{"a": 1, "b": 2}'::jsonb #- '{a}' => jsonb`, async () => {
 test("|/ 16 => number", async () => {
   await testQuery({
     query: `SELECT |/ 16`,
-    expected: [["?column?", { kind: "type", value: "number", type: "int4" }]],
+    expected: [["?column?", { kind: "type", value: "number", type: "float8" }]],
   });
 });
 
 test("||/ 27 => number", async () => {
   await testQuery({
     query: `SELECT ||/ 27`,
-    expected: [["?column?", { kind: "type", value: "number", type: "int4" }]],
+    expected: [["?column?", { kind: "type", value: "number", type: "float8" }]],
   });
 });
